// backup/client/src/types/User.ts
export interface User {
  emailAdress: string;
  companyLogo: string;
  aboutCompany: string;
  companyName: string;
  phoneNumber: string;
  recruiterName: string;
  webColor1: string;
  webColor2: string;
  id: string;
  token: string;
}
export type UserContextType = {
  emailAdress: string;
  companyLogo: string;
  aboutCompany: string;
  companyName: string;
  phoneNumber: string;
  recruiterName: string;
  webColor1: string;
  webColor2: string;
  id: string;
  token: string;
};